refactor(RadioGroup): move to typescript
